# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "N8N AI Workflow Automation Platform API",
        "description": "Ultra-scalable workflow automation platform with real-time execution, multi-tenancy, advanced queuing system, and comprehensive API documentation.",
        "contact": {
            "name": "N8N AI Support",
            "email": "api@n8n-ai-lara.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Production API Server"
        },
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/queue/status": {
            "get": {
                "tags": [
                    "Queue",
                    "Monitoring"
                ],
                "summary": "Get Queue Status",
                "description": "Get comprehensive status of the background job queue system",
                "operationId": "getQueueStatus",
                "responses": {
                    "200": {
                        "description": "Queue status retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/QueueStatus"
                                        },
                                        "health": {
                                            "$ref": "#/components/schemas/HealthStatus"
                                        },
                                        "timestamp": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - Admin access required",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Access denied"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "QueueStatus": {
                "title": "Queue Status",
                "description": "Background job queue status",
                "properties": {
                    "default": {
                        "properties": {
                            "pending": {
                                "type": "integer",
                                "example": 5
                            },
                            "processing": {
                                "type": "integer",
                                "example": 2
                            },
                            "failed": {
                                "type": "integer",
                                "example": 1
                            }
                        },
                        "type": "object"
                    },
                    "high-priority": {
                        "properties": {
                            "pending": {
                                "type": "integer",
                                "example": 2
                            },
                            "processing": {
                                "type": "integer",
                                "example": 1
                            },
                            "failed": {
                                "type": "integer",
                                "example": 0
                            }
                        },
                        "type": "object"
                    },
                    "total": {
                        "properties": {
                            "pending": {
                                "type": "integer",
                                "example": 7
                            },
                            "processing": {
                                "type": "integer",
                                "example": 3
                            },
                            "failed": {
                                "type": "integer",
                                "example": 1
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "HealthStatus": {
                "title": "Health Status",
                "description": "System health monitoring",
                "properties": {
                    "overall_health": {
                        "description": "Health score 0-100",
                        "type": "integer",
                        "example": 85
                    },
                    "queues": {
                        "$ref": "#/components/schemas/QueueStatus"
                    },
                    "recommendations": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "High queue backlog detected",
                            "Consider scaling up workers"
                        ]
                    },
                    "last_checked": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Laravel Sanctum Bearer Token Authentication. Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}